Update script.js
This is the code I've been working on all Sprint. It is different from the Master Branch in that everything can be used in browser, there are no async functions, and several other difference. This is the backend work for the DIN test, only the DIN test was touched from the original code before the Sprint.