Update index.html
Edited the textbox so not it does not allow for user input, but the user can still change the value using the step buttons.

Added a "I Can Still Hear It" button which allows the user to proceed through the test automatically at each stage and properly increments the frequency to the next suggested frequency. The user still has the ability to manually change the frequency with the step buttons if the button was clicked on accident or for whatever reason.