Update index.html
Front end for the rest of John's and Sanal's concerns.

Now, simply pressing "No, I can't hear it" will no longer take you to the results page. Instead, the test narrows in on the user's current frequency and pinpoints a closer approximation by 100 intervals instead of 500-1000.

For example:

User stops hearing the tone at 5000 Hz. The test then sets the user up with another test, ranging from 3000 Hz to the 5000 Hz when they clicked the sound. The increase and decrease buttons do not go past this range and only incremement at 100 steps. Once the user is past 6000 Hz, then the test will instead guide them down their last 1000 Hz, to pinpoint a better approximation for what they can hear.

There is also a failsafe, so if the user accidentally clicks "No, I can't hear it" at 10,000 Hz for example, but they last were able to hear at 5000 Hz, then the test will reset at 9000-10000 Hz, then if they click it again, 8000-9000, and so on until the user no long clicks "No I can't hear it" on the lowest end of the range or the user eventually reaches 2000 and clicks it.